pyproject.toml
[project]
name = "django_hw"
version = "0.0.1"
dependencies = [
    "django==5.0.4",
    "psycopg==3.1.18",
    "python-dotenv==1.0.1",
    "djangorestframework==3.15.1",
    "drf-yasg",
    "psycopg2-binary",
    "gunicorn",
    "django-browser-reload"
]


[project.optional-dependencies]
dev = ["flake8", "black", "flake8-quotes", "pep8-naming", "isort", "coverage"]

[tool.black]
line-length = 79
target-version = ["py312"]


[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "django_hw.settings"
# -- recommended but optional:
python_files = ["test_*.py", "*_test.py"]
